diffusion: consolidate the model package from 17 files to 7
Adding a second model meant adding another 17 files, most of them 100-200
lines. Group them by role instead, so a model is seven files:

  arch.py          architecture config
  dit.py           the network
  components.py    both VAEs, text encoder, weight loading
  layout.py        geometry, packed sequence, patchify, initial latents
  conditioning.py  keyframes, references, noise aug, presentation
  denoise.py       the loop and its rectified-flow sampler
  pipeline.py      the 8 stages and the pipeline

Mechanical: bodies move unchanged, imports are unioned, callers are rewritten.
No behaviour change and no new abstractions.

The alternative -- promoting the generic-looking pieces to a shared layer
ahead of a second model -- does not survive inspection. Weight sharding sits
in the same file as H3's grouped-QKV reorder, patchify beside H3's packed
scatter, rectified-flow Euler under minimax_h3_* names. Each would have to be
split first, and the result would be a shared API with exactly one caller.
They graduate when a second model uses them.

Also drops the vestigial config surface: ComponentConfig, the components list
and tp_size were parsed and never read (the H3 loader hardcodes its subfolders
and the DiT runs at tp_size 1), so a second model author would have had to
reason about all three for nothing. ulysses_degree now simply has to equal
num_gpus, which is what the code already assumed.

Comment density is down to 0.21 prose lines per code line, the lowest of any
ATOM subsystem (model_engine 0.34, models and model_ops 0.26).

253 diffusion tests pass; the five that covered the removed config surface are
gone with it.
