+++ atom/diffusion/__init__.py
@@ -3,21 +3,16 @@
 
 """Diffusion (video/audio generation) subsystem for ATOM.
 
-A sibling of ``atom.model_engine``, not an extension of it: no KV cache, no
-prefill/decode split, no token-level batching. One request is one job of N
-denoise steps over several heterogeneous components, taking minutes, and
-parallelised by splitting a *single* request's sequence across every GPU
-(Ulysses) rather than sharding a replicated batch.
-
-Reused from the LLM side: ``atom.model_ops``, ``atom.model_loader``,
-``atom.utils.distributed``, and the ZMQ multi-process pattern.
+A sibling of ``atom.model_engine``, not an extension: no KV cache, no
+prefill/decode split, no token-level batching. One request is one minutes-long
+job of N denoise steps over several heterogeneous components, parallelised by
+splitting a *single* request's sequence across every GPU (Ulysses).
 """
 
-from atom.diffusion.config import ComponentConfig, DiffusionConfig
+from atom.diffusion.config import DiffusionConfig
 from atom.diffusion.request import DiffusionJob, JobStatus
 
 __all__ = [
-    "ComponentConfig",
     "DiffusionConfig",
     "DiffusionJob",
     "JobStatus",
